version 1.1 but wthiout more dao func and without dict

+++ app/src/main/java/com/example/mmmsssmmm/data/dictionary/ModelDictEntity.kt
@@ -25,7 +25,7 @@ import androidx.room.PrimaryKey
 
 data class ModelDictEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val brandId: Int,
+    val brandId: Long,
     val bodyTypeId: Int,
     val name: String,
     val yearStart: Int,
